+++ tatva/element/__init__.py
@@ -15,7 +15,9 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with tatva.  If not, see <https://www.gnu.org/licenses/>.
 
-from ._element import Element as Element
 
-from ._element import Line2, Tri3, Quad4, Tetrahedron4, Hexahedron8
+# pyright: reportUnusedImport=false
+
+from .base import Element as Element
+from .base import Hexahedron8, Line2, Quad4, Tetrahedron4, Tri3
 from .beam import BeamElement, HermiteElement, CorotationalHermiteElement, LinearElement